pyproject.toml
[project]
name = "cr_rule_simplification"
version = "0.1.0"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10,<3.11"
#license = { file = "LICENSE.txt" }
authors = [
    { name = "Cyber-Physical Systems Group - Technical University of Munich", email = "commonroad@lists.lrz.de" }
]
description = "A Toolbox for Simplifying Traffic Rules in CommonRoad Scenarios"
keywords = [
    "autonomous vehicles",
    "automated driving",
    "specifications",
    "temporal logic",
]
dependencies = [
    "commonroad-io>=2024.2",
    "commonroad-drivability-checker>=2024.1",
    "commonroad-route-planner>=2024.1, <2024.2",
    "numpy>=1.22.0, <2.0",
    "more-itertools~=10.5.0",
]

classifiers = [
    "Programming Language :: C++",
    "Programming Language :: Python :: 3.10",
    "Operating System :: POSIX :: Linux",
]

#[project.urls]
#Homepage = "https://commonroad.in.tum.de/tools/commonroad-reach"
#Documentation = "https://commonroad.in.tum.de/docs/commonroad-reach/"
#Forum = "https://commonroad.in.tum.de/forum/c/commonroad-reach/19"
#Source = "https://gitlab.lrz.de/tum-cps/commonroad-reach"

[project.optional-dependencies]
test = [
    "pytest>=6.2.2",
]
#docs = [
#    "Sphinx>=4.5.0",
#    "sphinx-autodoc-typehints>=1.18.1",
#    "m2r2>=0.3.0",
#]
build = [
    "scikit-build-core~=0.8.0",
    "nanobind~=2.0.0",
    "pathspec>=0.12.1",
    "pyproject-metadata>=0.7.1",
    "typing_extensions~=4.12.2",
    "cmake>=3.24",
    "commonroad-drivability-checker>=2024.1", # Required to load the Python module for stub generation
]
dev = [
    "pre-commit>=3.7.0",
]

[build-system]
requires = [
    "scikit-build-core~=0.10.7",
    "nanobind~=2.2.0",
    "typing_extensions~=4.12.2",
    "cmake>=3.24",
    "commonroad-drivability-checker>=2024.1", # Required to load the Python module for stub generation
]
build-backend = "scikit_build_core.build"

[tool.scikit-build]
build-dir = "build/{build_type}_{wheel_tag}"

minimum-version = "0.8"
cmake.version = ">=3.24"

[tool.cibuildwheel]
skip = ["pp*", "*-musllinux*"]
archs = ["x86_64"]
environment = { PIP_DISABLE_PIP_VERSION_CHECK = "1" }
build-verbosity = 1
test-extras = ["test"]

# Formatting
[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 120

# flake8 linting defined in .flake8
